A script to create Deb and RPM packages
use fpm[1] to create either rpm or deb packages, based on the platform
we're working on. This is overridable by the environment variable PKG.

This pckage contains both TTF and OTF fonts, as well as the
documentation. Currently the version needs to be bumped. It would be
better if that was read from metadata.